fix FoodController

+++ src/main/java/com/example/demo/controller/FoodController.java
@@ -1,16 +1,14 @@
 package com.example.demo.controller;
 
 import com.example.demo.model.Food;
-import com.example.demo.model.Message;
-import com.example.demo.service.extend.IFoodOrderService;
 import com.example.demo.service.extend.IFoodService;
-import com.example.demo.service.impl.MessageService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @CrossOrigin("*")
@@ -22,25 +20,43 @@ public class FoodController {
     private IFoodService foodService;
 
     @GetMapping
-    public ResponseEntity<List<Food>> findAll(){
-        return new ResponseEntity<>(foodService.findAll(), HttpStatus.OK);
+    public ResponseEntity<List<Food>> findAll() {
+        List<Food> foodList = foodService.findAll();
+        if (foodList.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(foodList, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Food> findById(@PathVariable Long id){
-        return new ResponseEntity<>(foodService.findById(id).get(),HttpStatus.OK);
+    public ResponseEntity<Food> findById(@PathVariable Long id) {
+        Optional<Food> foodOptional = foodService.findById(id);
+        if (!foodOptional.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(foodOptional.get(), HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<Food> save(@RequestBody Food food){
-        return new ResponseEntity<>(foodService.save(food),HttpStatus.CREATED);
+    public ResponseEntity<Food> save(@RequestBody Food food) {
+        return new ResponseEntity<>(foodService.save(food), HttpStatus.CREATED);
     }
+
     @PutMapping
-    public ResponseEntity<Food> update( @RequestBody Food food){
-        return new ResponseEntity<>(foodService.save(food),HttpStatus.OK);
+    public ResponseEntity<Food> update(@RequestBody Food food) {
+        Optional<Food> foodOptional = foodService.findById(food.getId());
+        if (!foodOptional.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(foodService.save(food), HttpStatus.OK);
     }
+
     @DeleteMapping("/{id}")
     public ResponseEntity<?> remove(@PathVariable Long id) {
+        Optional<Food> foodOptional = foodService.findById(id);
+        if (!foodOptional.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         foodService.remove(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
